healthconnect-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt] 
python-multipart
secure-smtplib
python-dotenv
requests==2.31.0
pyjwt==2.8.0
cryptography==41.0.4
